Remove redundant check for empty graph file
readGraph() inside the HashGeneration class no longer checks whether graphs (input file type 1) are empty, because this type of check was added to the FileUtility class' readFile() function in commit 20db813. The function checking this was stored inside AlgorithmUtility and was not used anywhere else, so it was removed entirely.