Improve input file parsing
The univocity-parsers library is now used to parse input files. The delimiter is now detected automatically, so both csv and tsv files are now accepted.
This changed the way input files are read and checked. GraphParameter.FILE now stores file objects instead of strings. 
Content/existence is checked via file.io (checkNotEmptyFileContent() was removed).
Old checks for whether or not a file is supplied via the command line must still be performed (== null) to avoid NullPointer exceptions when a file.io function is called on null.