Fix regression error that prevented gspan output
The visualisation method MotifToJsonConvertsion.convertMotif caused a null pointer exception error when running the analysis with the gspan algorithm because the hash map MiningState.supportedMotifsDFScode is not normally used in this mode.
This commit copies all frequent motifs encountered during the gspan mining step and stores them as DFScode objects in this hash map.